# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certificate Manager API",
        "description": "API REST unificada (v1) para la gestión completa de solicitudes de certificados digitales: creación, archivos adjuntos, emisión agnóstica de proveedor (correo legacy / Viafirma RA PKCS#10 Zero-Touch), pagos WOMPI, cuotas POSTPAID, analíticas IA, notificaciones de vencimiento y webhooks. Requiere autenticación OAuth 2.0 con Laravel Passport.",
        "contact": {
            "name": "Soporte Matias",
            "email": "soporte@matias.com.co"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 — única versión soportada"
        }
    ],
    "paths": {
...
        "/admin/quotas": {
            "get": {
                "tags": [
                    "Cupos Admin"
                ],
                "summary": "Listar todos los cupos (admin)",
                "description": "GET /admin/quotas — Listar todos los cupos",
                "operationId": "9a92c1f75ebbf22d954da20c23d8bd2c",
                "responses": {
                    "200": {
                        "description": "Lista paginada de cupos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No es administrador"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cupos Admin"
                ],
                "summary": "Asignar cupo POSTPAID a una empresa",
                "description": "POST /admin/quotas — Asignar cupo POSTPAID a empresa",
                "operationId": "f59beb9ab9d6700331213d82b396c05a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "quantity",
                                    "period_start",
                                    "period_end"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 50
                                    },
                                    "period_start": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-01"
                                    },
                                    "period_end": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-31"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Cupo mensual mayo 2026",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cupo asignado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CertificateQuota"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No es administrador"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/quotas/{id}": {
            "get": {
                "tags": [
                    "Cupos Admin"
                ],
                "summary": "Ver detalle de un cupo",
                "description": "GET /admin/quotas/{id} — Ver detalle de un cupo",
                "operationId": "4df57555f5a2be420250ce356362e958",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del cupo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CertificateQuota"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cupo no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/quotas/company/{id}": {
            "get": {
                "tags": [
                    "Cupos Admin"
                ],
                "summary": "Estado y historial de cupos de una empresa",
                "description": "GET /admin/quotas/company/{id} — Estado de cupos de una empresa",
                "operationId": "c72d0215ee321cd2648eef36756ccee7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de cupos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "status": {
                                                    "properties": {
                                                        "allocated": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "used": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "remaining": {
                                                            "type": "integer",
                                                            "example": 38
                                                        },
                                                        "expires_at": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "history": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CertificateQuota"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CertificateQuota": {
                "description": "Cupo POSTPAID asignado a una empresa por el admin LOPEZSOFT",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "company_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "allocated_quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "used_quantity": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 12
                    },
                    "remaining": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 38
                    },
                    "period_start": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-01"
                    },
                    "period_end": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-31"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "EXHAUSTED",
                            "EXPIRED"
                        ],
                        "example": "ACTIVE"
                    },
                    "billing_type": {
                        "type": "string",
                        "enum": [
                            "POSTPAID"
                        ],
                        "example": "POSTPAID"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Cupo mensual mayo 2026",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de acceso OAuth 2.0 obtenido desde /authentication/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}